encode valueReference as Long

+++ library/FMI4j/fmi-import/src/main/kotlin/no/mechatronics/sfi/fmi4j/common/types.kt
@@ -26,7 +26,7 @@ package no.mechatronics.sfi.fmi4j.common
 
 
 typealias Real = Double
-typealias ValueReference = Int
-typealias ValueReferences = IntArray
+typealias ValueReference = Long
+typealias ValueReferences = LongArray
 typealias RealArray = DoubleArray
 typealias StringArray = Array<String>
